BaseDao.java int id change to long id CommentDao.java int id change to long id TaskDao.java delete all methods TaskDaoImpl.java all methods was implemented. They are work with sout, but now methods getAll and getTaskById don't work. I don't know why. Task.java int id change to long id Status.java toString method and a few updates to enum Main.java I delete all rubbish

+++ src/main/java/com/jc/tm/database/dao/CommentDao.java
@@ -12,7 +12,7 @@ public interface CommentDao extends BaseDao<Comment> {
     void update(Comment comment);
 
     @Override
-    Comment getById(int id);
+    Comment getById(Long id);
 
     @Override
     List<Comment> getAll();
